+++ yearzero/YZRoll.js
@@ -73,7 +73,7 @@ class YZRoll {
 	}
 
 	/**
-	 * Wether the roll was pushed or not.
+	 * Whether the roll was pushed or not.
 	 * @type {boolean}
 	 * @readonly
 	 */
@@ -526,6 +526,24 @@ class YZRoll {
 		out.unshift(`<${this.game}>`);
 		return out.join(' ');
 	}
+
+	/**
+	 * Returns only the values of the dice in the roll.
+	 * @returns {string}
+	 */
+	toValues() {
+		return this.dice
+			.map(d => d.valueOf())
+			.join(', ');
+	}
+
+	/**
+	 * Returns the primitive value of the roll.
+	 * @returns {number}
+	 */
+	valueOf() {
+		return this.sum();
+	}
 }
 
 module.exports = YZRoll;
@@ -593,7 +611,7 @@ class YZDie {
 	}
 
 	/**
-	 * Wether this die has been pushed.
+	 * Whether this die has been pushed.
 	 * @type {boolean}
 	 * @readonly
 	 */
@@ -602,7 +620,7 @@ class YZDie {
 	}
 
 	/**
-	 * Wether the die can be pushed (according to its type).
+	 * Whether the die can be pushed (according to its type).
 	 * @type {boolean}
 	 * @readonly
 	 */
@@ -637,7 +655,7 @@ class YZDie {
 
 	/**
 	 * Pushes the die, according to its type.
-	 * @returns {number} The result, wether it has been pushed or not.
+	 * @returns {number} The result, Whether it has been pushed or not.
 	 */
 	push() {
 		this.previousResults.push(this.result);
@@ -682,5 +700,9 @@ class YZDie {
 	toString() {
 		if (this.type === 'modifier') return `${this.operator}${this.result}`;
 		else return `${this.operator}d${this.range}${this.type ? `[${this.type}]` : ''}`;
+	}
+
+	valueOf() {
+		return +`${this.operator}${this.result}`;
 	}
 }